tests/sriov_kvm: update negative input command
Reference to DPDK-3807: rule-id can be any value which is set by user.
So changed the invalid rule-id as 256 to test negative input commands.
Because in dpdk the mirror rule-id is uint8_t.

Remove command "set port 0 mirror-rule 2 vlan-mirror 9 dst-pool 1 on".
It's not a invalid command for vlan-mirror is supported now.

Signed-off-by: Fangfangx Wei <fangfangx.wei@intel.com>