tests/pmdrss_hash: set next header for ipv6-sctp packet
The rss type of the flow type ipv6-sctp should be set as sctp instead of ip.
And it need to explicitly set the next header nh=132 for scapy to send ipv6
sctp packets. Otherwise, the scapy below the version 2.3.2 will send ipv6
sctp packets with unknown next header(nh=59).

Signed-off-by: Luo Gaoliang <michael.luo@intel.com>