framework/pktgen: add configuration for start T-Rex during preparation
The currently merged patchset for T-Rex does not actually start T-Rex.
Add a configuration option "start_trex" and start T-Rex during the
prepare_generator() step if this is set to a non-empty string.  We need
T-Rex running during this step so we can determine which ports the T-Rex
API is making available to the DUT before we set up the DUT's network
interfaces accordingly.

This configuration option assumes that T-Rex will always be run on the
tester node.

Signed-off-by: Patrick MacArthur <pmacarth@iol.unh.edu>