tests/nic_single_core_perf: report results in JSON format
Add a JSON report file in addition to the texttable report file, for
easier consumption by other automated tools. The JSON output produced
will look like:

    {
      "results": [
        {
          "parameters": { "frame_size": 64, "txd/rxd": 1024 },
          "throughput": { "result": "PASS", "delta": -0.452, "unit": "Mpps" }
        },
        /* ... */
      ]
    }

Each entry in the results list is essentially a table row. The
parameters "frame_size" and "txd/rxd" are the input parameters for
each given measurement. Each other element in the result dictionary
is an output measurement for the given measurement.

Signed-off-by: Patrick MacArthur <pmacarth@iol.unh.edu>