framework/crb: Add new option in crbs to identify dut arch
The SOCKET IDs on some arm64 servers are incorrect which listed by lscpu
on Linux. This defect will causes DTS failures.

0,0,0,0
1,1,0,0
2,2,0,0
3,3,0,0
4,4,1,0
5,5,1,0
6,6,1,0
7,7,1,0
8,8,2,0
9,9,2,0
10,10,2,0
11,11,2,0
12,12,3,0
13,13,3,0
14,14,3,0
15,15,3,0
16,16,4,1
17,17,4,1
18,18,4,1
19,19,4,1
20,20,5,1
21,21,5,1
22,22,5,1
23,23,5,1
...

However, the NUMA NODE ID is correct and the NUMA NODE ID is the
equivalence of SOCKET ID on arm server.
By replacing the SOCKET ID with the NUMA NODE ID on the arm64
platform can solve this issue.

Signed-off-by: Phil Yang <phil.yang@arm.com>
Tested-by: Liang JiangWang <liangjiang.wang@arm.com>