framework/packet: support packet sniffer to specify running machine
By default, the Tester is supposed to be the server which running DTS
and the packet sniff methods are running on it.
However, if DTS was not running on the Tester and the Tester is another
remote server, so packet sniff methods cannot sniff Tester's packets
correctly.

This patch adds support for packet sniff methods to specify running
machine and implements sniff packet methods in class tester.

1. Add parameter to sniff_packet and load_sniff_pcap methods to specify
the running machine.
2. Remove load_sniff_packets method in packet.py.
3. Add tcpdump_sniff_packets, load_tcpdump_sniff_pcap and
load_tcpdump_sniff_packets for class tester with the new sniff_packet
API.
4. Update tester.check_random_pkts method with
tester.tcpdump_sniff_packets and tester.load_tcpdump_sniff_packets to
make it execution on the Tester.

Signed-off-by: Phil Yang <phil.yang@arm.com>
Suggested-by: Marvin Liu <yong.liu@intel.com>
Suggested-by: Tu Lijuan <lijuan.tu@intel.com>