test/kni/ping add a -m parameter when start
Due to changes in dpdk, now by default kni interface carrier status is
"off". If link monitoring is enabled with the -m command line flag,check
the link status of each physical NIC port and will Update the carrier
status of the corresponding KNI interface to match the status of the
physical NIC port's state.

Signed-off-by: zhuwenhui <wenhuix.zhu@intel.com>