framework/dut.py: pull the name after bound the interface to kernel driver
When restoring the interface, it will do unbinding, then bind the interface
with the kernel driver, it's better to pull the interface name with 1s interval.
Because on some system, especially on VMs, if try to get the name immediately
after bind the interface with the driver, it may not be ready in its sysfs.

Signed-off-by: Luo Gaoliang <michael.luo@intel.com>