tests/coremask: correct Master lcore
The "test_all_cores_core-mask sub-test validates that the Master lcore
becomes ready, the test assumes this to be core 1, this is not correct
for all architectures or all configurations.  For example this test will
fail on x86_64 with bypass_core0=False and on Power9 with most smt
configurations.  This patch determines the first (lowest numbered) core
from the list of available cores and tests that that core becomes ready.

I validated this change that the test will pass on both Power9 and
x86_64 with both true and false setting of bypass_core0.o

Signed-off-by: David Wilder <dwilder@us.ibm.com>