framework: add a new subtitle paramater
Add a new subtitle paramater to main. There are use cases when we want
to add specific details to the rst report, for example, FW version,
name of the relevant project, setup details, etc. This patch provides this
ability. With this patch, you can run, for exampe:

./dts -s --config-file=execution.cfg --subtitle=FW_1.8.0

to add FW_1.8.0 in the rst report.

Signed-off-by: Rami Rosen <ramirose@gmail.com>