tests/vf_kernel: Optimize validation results and scripts
1. Modify the verification result of the case.
  Before each case runs, we need to check that the connection state of the
corresponding NIC is yes. If no, we need to call the restore_interfaces_linux()
method to restore the NIC state to yes, and there is no need to do any other
operations.

2.Add code to check NIC status.
  Sometimes, ifconfig %s up does not take effect due to the case running
too fast, and it needs to wait for a certain period of time.

3.Add time.sleep() to enable tcpdump to start normally.
  Sometimes, the tcpdump command requires a certain open time, otherwise
othe corresponding packet cannot be fetched.

4.Replace the virtual NIC intf0 to intf1.
  When all cases run together, it will affect the intf0 NIC required by
test_vlan, so replace intf0 to intf1.

Signed-off-by: zhang,yan <yanx.a.zhang@intel.com>