framework/packet: remove the set of conf.use_pcap
when set scapy conf.use_pcap=True,
in scapy 2.2 and 2.3, it will affect the sniff and rdpcap method,
it will use the libpcap to read packet.
in scapy 2.4, it will affect the sendp method, it will use module
in pcapy, so you should upgrade pcapy to 0.11.3 if you want to set
the conf.use_pcap to True.
But after try, find it is not necessary to set the conf.use_pcap to
True, it no effect to our cases. so remove the config in the file,
that we no need to do anything after upgrade scapy.

Signed-off-by: lihong <lihongx.ma@intel.com>