tests/TestSuite_mac_filter.py Correction in while loop condition.
Removing '=' from while loop condition as i is initialized to 0 and the loop
iterates upto 256 resulting in total 257 iterations, as a result of that an
extra unwanted invalid mac address is added at the end i.e.
"00:01:00:00:00:100",it returns not expected String which doesn't match with
the string in verify function, Leading to test failure.

Signed-off-by: yash <ysharma@marvell.com>