tests/shutdown_api:'port stop' recover the orignal link status for CVL
In DPDK fdcf92ed6637b790, the 'port stop' will recovery the oringnal link
status for CVL.
If the link is up, it should be up after 'port stop'; if down, then be down.
This is different with other NICs.

In test case stop_restart, the port is up before 'port stop', so verify
the port is UP for CVL and down for other NICs.

Signed-off-by: Wenjie Li <wenjiex.a.li@intel.com>