tests/TestSuite_vf_jumboframe:Rx packet size need add 4 bytes
DTS framework has switched i40evf to iavf, so align with it.
After testing, found that:
when send a packet which length is 1513 bytes.
for net_iavf driver; the statistic from "show port stats 0" is: RX-bytes: 1513
for net_i40e_vf driver; the statistic from "show port stats 0" is: RX-bytes: 1517

Signed-off-by: Xie wei <weix.xie@intel.com>