tests/TestSuite_veb_switch:vf1 RX-bytes need add 4 bytes
DTS framework has switched i40evf to iavf, so align with it.
After testing, found that:
when send a packet which length is 60 bytes.
for net_iavf driver; the statistic from "show port stats 0" is: RX-bytes:56,
but set verbose 1, we can see the packet length is 60.
for net_i40e_vf driver; the statistic from "show port stats 0" is: RX-bytes: 60

Signed-off-by: Xie wei <weix.xie@intel.com>