meson compile: Improve compile times
Increased compile timeout
Removed workers parameter from ninja compilation command

After a review of CI metrics, we noticed that dts runtimes had increased
greatly over the last few months. A review of recent logs showed
timeouts for compiling dpdk. Upon review, we discovered that, if the
"dut_cores" parameter is not set in conf/crbs.cfg, then dts uses the
default value, which is a single core, as the core list. The number of cores
in this list was then used the number of workers for ninja. This causes DTS
to attempt to compile DPDK serially, resulting in greatly increased compile
times. The IOL's development machine saw a 30x compile speed increase from
the included changes.

Signed-off-by: Owen Hilyard <ohilyard@iol.unh.edu>