bugfix: Fixed a timeout with mlx5_core on ARM
The time to unbind this driver has been increasing with DPDK releases, to
the point that it is now about 14 seconds. This means that, with the
default timeout of 15 seconds, there isn't much room for any kind of
disruption during testing. DTS throwing a timeout error here is not
indicative of an acutal problem, and as such I have increased the timeout
to allow continued stable functioning. The new 30 second timeout should
have minimal affect on the detection speed of any real error.

Signed-off-by: Owen Hilyard <ohilyard@iol.unh.edu>