tests/nic_single_core_perf: revise rx desc size
rx_desc_16byte uses  'y' indicates the rx descriptor size to 16B, and
other string for 32B. but 'y' is a reserved string in INI, misuse it
will cause the write-back configuration file can't be read correctly.
so rename rx_desc_16byte to rx_desc_size and the type is integer.

Signed-off-by: Lijuan Tu <lijuan.tu@intel.com>