Strip prompt by default in send_expect
During investigation of timeout exceptions during performance testing, the
community lab found that trailing whitespace in run commands was the cause
of some of the timeouts. This patch makes the Crb version of send_expect
strip whitespace from the expected string by default. It is possible to
disable this behavior should it be needed, but we expect that the majority
of use-cases are matching non-whitespace characters.

Signed-off-by: Owen Hilyard <ohilyard@iol.unh.edu>