framework/config: fix issue with string values
String values were ran through 'eval' like everything else as part of
parsing config files. This will work the first time if there are quotes
around the value, but if the --update-expected flag is passed, it will
re-write the config file without quotes around the string value. This
then cause an error when the value is 'eval'ed again, since it will now
be evaluated as a variable.

Signed-off-by: Owen Hilyard <ohilyard@iol.unh.edu>