framework/settings: Add broadcom nics to dts
brcm_57414 and brcm_P2100G are not currently in the settings files for
DTS. This causes errors when DTS is run with these nics. The community
lab was running a patched version of DTS, but we are making an effort to
more closely align to upstream. Since these nics are used in CI for
DPDK, they are well tested with DTS and shouldn't cause any issues.

Signed-off-by: Owen Hilyard <ohilyard@iol.unh.edu>