tests/power_brach_ratio: refactor power test
This patch will make the test run on host instead of vm.
Since vm is not required all vm code is removed.

before test was comparing min max freq values to P1 which
was not what we wanted, also the logical operator used didn't
allow for some variance and forced to be exact freq number.
So we fixed that by comparing current freq value of the core
to P1 while in traffic and Logical operator have been changed
to allow some variance when comparing with P1

Some addition to both .cfg and .py value for clear debugging
example you can specify the range(from to) for vm power mgr
and testpmd, also you can specify which exact core you want
to check for test.

Signed-off-by: Harneet Singh <harneet.singh@intel.com>
Tested-by: Reshma Pattan <reshma.pattan@intel.com>