tests/cvl_dcf_switch_filter: modify the mask value in testsuite
the ipv4 pattern, when the action is drop, and the mask is not 255.255.255.255,
the rule created is as a acl rule. only the acl rule table is full, the rule is created as switch rule.
when the mask is 255.255.255.255, the rule is created as switch rule.
so modify all the mask to 255.255.255.255 when the action is drop.

Signed-off-by: Peng Yuan <yuan.peng@intel.com>
Tested-by: Qin Sun <qinx.sun@intel.com>