tests/cvl_switch_filter: delete the placeholder fdir rule
in non-pipeline mode, "priority" is ignored, a flow rule can be created as a flow director rule or
a switch rule depends on its pattern/action. if a rule is supported by switch or fdir at the same time,
it will be created in the switch table first. we no longer need to populate the fdir table first

dpdk commit: e4a0a7599d974f05665fec3e4c251659f0b11453

Signed-off-by: Jiale Song <songx.jiale@intel.com>