framework/flow: Remove items not supported by testpmd
Testpmd does not have full support for rte_flow at this point, and the
rte_flow test suite has not been kept up to date with changes in
rte_flow. As such, the logic for generating the flows had to be updated
to disable some test suites to avoid false failures.

Signed-off-by: Owen Hilyard <ohilyard@iol.unh.edu>