formatting_script: Add formatting script
Added a formatting script to run black an isort with. By default, it
will run in the current working directory. If $1 is specified and not
either "-h" or "--help", it will run in that directory instead. This
does not allow checking for a git diff after formatting (for rejecting
patches in CI), but that should be fairly easy to implement in a wrapper
script.

Signed-off-by: Owen Hilyard <ohilyard@iol.unh.edu>