ixia_network/ixnet: fix bug of changing string content
In history patch that add OUTPUT_DIR variable into initional function to
fix E1101 error that instance has no member error[pylint checking]. this
intruduce a bug, OUTPUT_DIR was defined in subclass as a property contained
a string  , and when parent class initial, the OUTPUT_DIR attribute was
already exist and couldn't redefine again. this patch attempt to fix it
that  through reflection mechanism to get the value instead of using
the variable directly and to avoid the E1101 error.

Signed-off-by: Jun Dong <junx.dong@intel.com>