tests/tso: improve checksum verify code
Reimplemented the checksum calculation with a non-file method.
Replaced property pks with a local variable pkts.
Renamed local variable pks to pkt as it stands for a single packet.
Use the name of layers as the index of packet to check if it is in a packet rather than check if it is in the dumped string of a packet.
Renamed checksum_list and checksum_list1 to chksum_list_rx and checksum_list_good.

Signed-off-by: Ke Xu <ke1.xu@intel.com>
Reviewed-by: Lijuan Tu <lijuan.tu@intel.com>