tests/ice_ecpri: optimize script to solve case random failed issue
fix 2 issue that cause the case to fail randomly:
    1. when the hash value is different, the queue may be the same. the number of different queues deleted is equal to
the number of different hash values.
    2. because the queue allocated by the packets hitting the fdir rule may be the same as the queue calculated by the
hash, but the test is to check that the queue conforming to this rule is different from the queue calculated by the hash
value, which causes the case test to fail. therefore, delete the hash value calculation check in this case.

Signed-off-by: Song Jiale <songx.jiale@intel.com>