tests/ice_dcf_switch_filter: update existing TO_VF action cases to cover port representor action from not covered vf original action
Deprecated TO_VF action contains a 'vf original' action which is not covered in previous tests. This patch applys rule to previous cases for 'vf id' action to cover 'vf original' action.

This commit added extra rules and configures to apply the vf original rules. This commit also modified implementation of several cases to support the check of vf original action and the check of multiple ports.

Signed-off-by: Ke Xu <ke1.xu@intel.com>
Acked-by: Lijuan Tu <lijuan.tu@intel.com>