vf_vlan: modify code to adapt dcf vlan offload
After sync with dev,change rx path is not safe on device.
Usually the device driver requires the device not be started to change offloads.
For now the test result under vector path is normal.
So we change to scalar path to test dcf vlan offload.

dcf don't support vlan on vector path, and change offloads are not
allowed when path is selected. So use specific scalar path to workaround
this issue.

Signed-off-by: Zhimin Huang <zhiminx.huang@intel.com>