framework/dut: add support for restoring dual-port NIC with single PCI address
The functions of restore_interfaces and rescan_ports get the incorrect
interface and mac for the second port of the NIC which has two ports but
single PCI address.

Add support for it by adding the filed of port_id for port_info, then it
can get correct interface and mac by distinguishing port_id.

Signed-off-by: Qin Ke <qin.ke@corigine.com>
Reviewed-by: Niklas Söderlund <niklas.soderlund@corigine.com>
Reviewed-by: Chaoyong He <chaoyong.he@corigine.com>
Signed-off-by: Niklas Söderlund <niklas.soderlund@corigine.com>
Reviewed-by: Lijuan Tu <lijuan.tu@intel.com>