tests/telemetry: Optimize the parameter for function json.load
In Python 3.10, the json.load() function removed the encoding parameter,
because in Python 3, the json module uses UTF-8 encoding by default to process JSON data.
As such, the encoding argument is no longer required, and has been removed in Python 3.10.
If you need to specify other encoding methods,
you can use the encoding parameter in the open() function to specify.
Please refer to the Python official documentation for more details.
https://docs.python.org/3/library/json.html

Signed-off-by: Dukai Yuan <dukaix.yuan@intel.com>
Acked-by: Lijuan Tu <lijuan.tu@intel.com>